Implement GenerateCredentials and RevokeCredentials in the ElastiCache provider
The GenerateCredentials call returns with the client credentials. This is the
same for all application bound to the same service instance.

The RevokeCredentials does nothing as we are not able to change the replication
group password or even handle multiple credentials.

The auth token generation was moved from the broker to the provider as the
broker should not know about authentication details.