Delete cache parameter group after the replication group was deleted
We can create only a limited number of cache parameter groups on AWS therefore we have to clean up the ones we don't need anymore. Also we don't want to leave unnecessary resources in our AWS accounts.
The cache parameter group can not be deleted while the replication group still exists and the deletion usually takes a long time. We decided to do the clean up in the last operation API call.

We return with operation specific data in the provision/deprovision/update calls so we can identify the last action in the last operation API call (CF will pass back this operation data to the broker).

If deleting the cache parameter group fails CF will keep calling the last operation endpoint until it succeeds.