Configure CacheParameterGroup from plan config
Previously we had one elasticache parameter group per instance. This was
because there are several useful configuration options that tenants may
have wanted to set to tune the behaviour of their elasticache instance.

Unfortunatly AWS is quite strict with the number of unique
CacheParameterGroups and does not give us a way to easily know when we
were hitting the limits.

To work around this limit we have decided to have one parameter group
per PLAN so that we can more easily predict the number of parameter
groups required. The downside to this is that important parameters
(suchas the memory eviction policy for redis) must now be configured at
plan-level so a larger number of plans are required to support the
various configurations that tenants may require.

This change requires that a `CacheParameterGroupName` be set in the
`PlanConfig` and that name will be used to create a parameter group if
it does not exist.