Use cluster-mode redis and enable backups
We want to be able to enable automatic backups.

Snapshot backups are only supported for t2.* node types when using redis
in "cluster mode", so we are only going to support cluster mode redis
going forward.

We change the behaviour of the Shards configuration parameter so that
even when set to "1" we use enable cluster mode.

We must always set AutomaticFailoverEnabled when using cluster mode so
we remove this as an option that can be set from the plan config.

We hardcode the default SnapshotWindow to a known default value so we can
document it.