Reorganise providers and broker code
Reorganise the code to separate the concerns in different modules.

We move the generic provider code to the modules `providers` and the
specific redis implementation to `providers/redis`

The Provider interface is still too specific and coupled to
the redis specifics, but we do not want to change the signatures
and types, only relocate them in different modules.