Provision an instance from another's latest snapshot
Implement the logic in the broker controller to accept a custom
parameter from the user on creation: restore_from_latest_snapshot_of

The user shall pass the GUID of the service instance from which
the snapshot has been created on:

  cf create-service ... -c { "restore_from_latest_snapshot_of": ""}

The broker would delegate on the specific provider to find
all the snapshots for a given service instance GUID, and
try to restore from the latest if any.

We extended the provider interface with the method FindSnapshots()

For security we only allow restore from snapshots from instances
from the same Org GUID, Space GUID, by checking the tags of
the snapshot that are automatically populated by AWS when
creating them.

We only allow to restore from snapshots with the same Plan GUID,
to ensure compatibility, although we are not aware of any