Show detailed status in last-operation
We want to be able to see as much detail of the state of the service as
possible. This is particulaly important since we now allow setting
configuration parameters on Provision/Update and we need a way to read
back what values have been configured.

The open service broker api does not provide a propper solution for this
so we are hijacking the "message" that is displayed within last
operation to dump a bunch of key/value messages as a workaround.

The message is a best-effort, and will only show details if/when it can
and fallback to less information when not available. The idea being that
it should never cause last operation to error just because we failed to
generate a pretty message.

To ensure that the LastOperation message is polled for after update we
return an IsSync: true respond after Update (even though it is not
strictly required in our case)