Implement CLI tool for generating cache cluster names
Cache cluster names in AWS are generated by hashing the service GUID from Cloud
Foundry. We tag each cluster with the service GUID, but the AWS console does
not show these tags anywhere, making it difficult to find the right cluster
given the service GUID. To assist with this, we implement a tool to hash a
given string using the same function as the broker uses, so that the AWS
console can be searched using the cluster name.