Remove unused CacheParameterGroupName field from ProvisionParameters struct
When creating an AWS Elasticache Cache Cluster, you need to provide
CacheParameterGroupName. In the Redis provider, the cache parameter group is
created and managed, and the name inserted. However, the name was still being
passed to the broker's Provision method, and ultimately ignored.

This commit removes it from the struct, so it doesn't need to be passed around.