Apply cost allocation tag to new service instances
We want to be able to see how much each service instance is
costing us. The best way to do that in the AWS bill is to apply
cost allocation tags to things. In this case, the tag is
'chargeable_entity' and we give it a value matching the service
instance id, so that we can see cost at a service instance
level.